training/requirements.txt
torch>=1.8.0
torchvision>=0.9.0
scikit-image>=0.19.0
matplotlib>=3.4.0
tensorboard>=2.7.0
numpy>=1.20.0
tqdm>=4.60.0
pillow>=8.2.0
scipy>=1.7.0
flask>=2.0.0 